apply (broken) algorithm v1

+++ packages/schemas/src/answer.ts
@@ -13,5 +13,5 @@ export interface Ops {
 }
 
 export interface AnswerResponse {
-  revision: number & tags.Type<"uint32">;
+  revision: number & tags.Type<"uint64">;
 }
